Updated topology and network metrics, and XML parser
Metrics now use latency, jitter, and packetloss on links instead of
CDFs. Built-in plugins were updated, and examples are included in
resource/. Real-life data from planetlab ping measurments is also
included in resource/ as a good source of topology data for plugins to
utilize.

+++ src/node/shd-network-interface.c
@@ -33,9 +33,9 @@ struct _NetworkInterface {
 	Network* network;
 	Address* address;
 
-	guint32 bwDownKiBps;
+	guint64 bwDownKiBps;
 	gdouble timePerByteDown;
-	guint32 bwUpKiBps;
+	guint64 bwUpKiBps;
 	gdouble timePerByteUp;
 
 	/* (protocol,port)-to-socket bindings */
@@ -58,7 +58,7 @@ struct _NetworkInterface {
 };
 
 NetworkInterface* networkinterface_new(Network* network, GQuark address, gchar* name,
-		guint32 bwDownKiBps, guint32 bwUpKiBps) {
+		guint64 bwDownKiBps, guint64 bwUpKiBps) {
 	NetworkInterface* interface = g_new0(NetworkInterface, 1);
 	MAGIC_INIT(interface);
